Implement the first specialised bits
This commit gives an idea of how the specialised module will be
structured. It is not finished yet.
We also use hashbrown's `HashSet` directly everywhere. Furthermore, I
changed `int` to be a `usize`; it's a complete mess now and has to be
made more consistent.

+++ src/lib.rs
@@ -34,7 +34,8 @@ pub mod fix_int {
     ///
     /// This type is generally used in this library to represent node/vertex labels in
     /// graphs (and related stuff).
-    pub type int = u32;
+    // pub type int = u32;
+    pub type int = usize;
 
     /// Run this to check that we are on a 32bit platform.
     pub fn ensure_at_least_32bit_platform() {
@@ -57,7 +58,8 @@ pub mod fix_int {
     macro_rules! enumerate {
         ($iter:expr) => {
             // always make sure that we actually use the int type here
-            (0u32..).zip($iter)
+            // (0u32..).zip($iter)
+            (0usize..).zip($iter)
         };
     }
     pub(crate) use enumerate;
